\name{secrpoly-package}
\alias{secrpoly-package}
\alias{secrpoly}

\docType{package}

\title{Area-search Spatially Explicit Capture--Recapture }

\description{

  Functions to estimate the density and size of a spatially
  distributed animal population sampled by searching polygons or transects.

}

\details{

\tabular{ll}{
Package: \tab secrpoly\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2025-11-19\cr
License: \tab GNU General Public License Version 2 or later\cr
}

This contains a subset of the functionality in \pkg{secr}. 
In future that may be removed from \pkg{secr}.

}

\author{ Murray Efford \email{murray.efford@otago.ac.nz} }

\references{

Borchers, D. L. and Efford, M. G. (2008) Spatially explicit maximum
likelihood methods for capture--recapture studies. \emph{Biometrics}
\bold{64}, 377--385.

Efford, M. G. (2011) Estimation of population density by spatially
explicit capture--recapture with area searches. \emph{Ecology}
\bold{92}, 2202--2207.

}

\keyword{ package }

\seealso{

\code{\link{secrpoly.fit}}

}

\examples{

\donttest{

predict(FTHL.0)

}
}